Split routes into a new folder
