+++ backend/src/routes/auth.js
@@ -1,0 +1,34 @@
+import { Router } from "express";
+import { AuthService } from "../services/auth.js";
+
+const router = Router();
+
+router.get("/signin", async (req, res) => {
+  console.log("Received request at /auth/signin:", req.query);
+  const { email, password } = req.query;
+  const result = await AuthService.signin(email, password);
+  if (result?.error) {
+    res.status(401);
+  }
+  console.log("Returning:", result);
+  res.send(result);
+});
+
+router.post("/signup", async (req, res) => {
+  console.log("Received request at /auth/signup:", req.body);
+  const params = req?.body?.params;
+  if (!params) {
+    res.status(400);
+    res.send("No parameters provided.");
+    return;
+  }
+  const { email, password, username } = params;
+  const result = await AuthService.signup(email, password, username);
+  if (result?.error) {
+    res.status(401);
+  }
+  console.log("Returning:", result);
+  res.send(result);
+});
+
+export const authRouter = router;
